Add Apache 2.0 License file
:Release Notes:
The Apache License version 2.0 file is added to the repository.

:Detailed Notes:
A new file named `Apache-2.0` has been added, which includes the full
text of the Apache License version 2.0. This license allows users to
use, reproduce, and distribute the work under specified conditions. It
also includes terms for patent rights and limitations of liability.

:Testing Performed:
Local Test: OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-15607] Fix Security Errors and Wno errors - pmlogdaemon

Change-Id: Ie879ed50a121bb933b374d28075336b1c4b6c81d