Rename license file
:Release Notes:
Renamed the license file from Apache-2.0 to LICENSE

:Detailed Notes:
The file containing the Apache-2.0 license has been renamed to LICENSE
to follow common naming conventions and improve clarity. This change
does not alter the content of the license, ensuring compliance and
consistency with the original terms.

:Testing Performed:
Local Test: OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-15607] Fix Security Errors and Wno errors - pmlogdaemon

Change-Id: I75c7f62c0d9268350845363ebf7d1095ff83a34e